added more maps; changed drawing mechanism
the first call to ondraw method it collects all static objects (walls)
and store a bit map with all of them in the correct position. The
following calls will use this map as a template to put the rest of the
objects that can change position during the game.

+++ src/ist/meic/bomberman/engine/DrawableObject.java
@@ -52,7 +52,7 @@ public abstract class DrawableObject {
 		return coord*MATRIX_TO_PIXEL;
 	}
 	
-	public int destroy(){
+	public double destroy(){
 		return 0;
 	}
 }
